@@ -2,15 +2,10 @@
   <div class="mt-10">
     <div class="d-flex justify-space-between">
       <span class="ml-3">Choose a survey</span>
-      <v-btn class="mr-3" color="green" @click="refreshSurveys">Refresh</v-btn>
+      <v-btn class="mr-3" color="green" text="Refresh" @click="refresh" />
     </div>
     <v-list dense fill-height fluid class="list-group">
-      <v-list-item
-        v-for="{ key, label, to } in surveyLinks"
-        :key="key"
-        :to="to"
-        class="list-group-item"
-      >
+      <v-list-item v-for="{ key, label, to } in surveyLinks" :key="key" :to="to" class="list-group-item">
         {{ label }}
       </v-list-item>
     </v-list>
@@ -19,38 +14,24 @@
 
 <script lang="ts">
 import { defineComponent } from "vue";
-import { mapState, mapActions } from "pinia";
-import { koordStore } from "../store";
-
-type SurveyLink = {
-  key: string | number;
-  label: string;
-  to: string;
-};
+import { storeToRefs } from "pinia";
+import { useSurveyStore } from "../store/surveyStore";
 
 export default defineComponent({
   name: "SurveyList",
-  props: { username: { type: String, required: true } },
+  setup() {
+    const store = useSurveyStore();
 
-  computed: {
-    ...mapState(koordStore, ["surveys", "getSurveys"]),
-    surveyLinks(): SurveyLink[] {
-      const surveyIds: number[] = [...this.getSurveys];
-      return surveyIds.sort().map((surveyId) => {
-        const title = this.surveys[surveyId].surveyls_title;
-        return {
-          key: surveyId,
-          label: `${surveyId} - ${title}`,
-          to: `/survey/${surveyId}`,
-        };
-      });
-    },
-  },
-  async mounted() {
-    await this.refreshSurveys();
-  },
-  methods: {
-    ...mapActions(koordStore, ["refreshSurveys"]),
+    const { surveyLinks } = storeToRefs(store);
+
+    function refresh() {
+      store.refreshSurveys();
+    }
+
+    return {
+      surveyLinks,
+      refresh,
+    };
   },
 });
 </script>
