@@ -2,16 +2,12 @@
   <v-container fluid class="px-3 mx-0">
     <v-row class="pt-3">
       <v-col cols="12" class="avoid-page-break px-1 py-1">
-        <display-options :display-options="showOptions" :options="options" />
+        <display-options :options="tableOptions" />
       </v-col>
     </v-row>
     <v-row class="pt-3">
       <v-col cols="12" class="avoid-page-break px-1 py-1">
-        <v-data-table
-          :items="filteredRecords"
-          :headers="headers"
-          no-data-text="There are no records to show"
-        >
+        <v-data-table :items="filteredRecords" :headers="headersFromKeys" no-data-text="There are no records to show">
           <template #item.token="{ item }">
             <span>
               {{ getParticipant(item.token) }}
@@ -25,12 +21,11 @@
 
 <script lang="ts">
 import { ResponseModel } from "../../types/response.model";
-import { ParticipantModel } from "../../types/participant.model";
 import DisplayOptions from "./DisplayOptions.vue";
-import { koordStore } from "../../store";
-import { defineComponent } from "vue";
-import { mapState } from "pinia";
+import { useSurveyStore } from "../../store/surveyStore";
+import { defineComponent, computed } from "vue";
 import { tableOptions } from "./options";
+import { getParticipant } from "../../helpers/chartFunctions";
 
 interface Header {
   title: string;
@@ -44,89 +39,46 @@ export default defineComponent({
   components: {
     DisplayOptions,
   },
-  props: {
-    qKeys: {
-      type: Array<string>,
-      default: () => [],
-    },
-    responses: {
-      type: Array<ResponseModel>,
-      default: () => [],
-    },
-    participants: {
-      type: Array<ParticipantModel>,
-      default: () => [],
-    },
-    showOptions: {
-      type: Boolean,
-      default: false,
-    },
-  },
-  data() {
-    return {
-      options: tableOptions,
-    };
-  },
-  computed: {
-    ...mapState(koordStore, ["settings"]),
-    filteredRecords(): ResponseModel[] {
-      return this.responses.filter(
-        (response: ResponseModel, index: number, array: ResponseModel[]) => {
-          if (this.settings.onlyActive) {
-            const token = response.token;
-            const lastResponse = array
-              .filter((item) => item.token === token)
-              .reduce((prev, current) => {
-                return new Date(prev.submitdate) > new Date(current.submitdate)
-                  ? prev
-                  : current;
-              });
-            return response === lastResponse;
-          }
-          return true;
+  setup() {
+    const store = useSurveyStore();
+
+    const filteredRecords = computed(() => {
+      return store.responsesInTimeline.filter((response: ResponseModel, index: number, array: ResponseModel[]) => {
+        if (store.settings.onlyActive) {
+          const token = response.token;
+          const lastResponse = array
+            .filter((item) => item.token === token)
+            .reduce((prev, current) => {
+              return new Date(prev.submitdate) > new Date(current.submitdate) ? prev : current;
+            });
+          return response === lastResponse;
         }
-      );
-    },
-    showKeys(): string[] {
-      const qk = this.qKeys;
-      qk.unshift("participant");
+        return true;
+      });
+    });
+
+    const showKeys = computed<string[]>(() => {
+      const qk = [...store.questionKeysWithSubquestions];
       qk.unshift("submitdate");
       qk.unshift("token");
       return qk;
-    },
-    headers(): Header[] {
-      const headers = this.responses.reduce(
-        (acc: Record<string, Header>, response: ResponseModel) => {
-          const keys = Object.keys(response);
-          keys.forEach((key) => {
-            if (!acc[key]) {
-              acc[key] = {
-                title: key,
-                key: key,
-                align: "start",
-                sortable: true,
-              };
-            }
-          });
-          return acc;
-        },
-        {}
-      );
-      return Object.values(headers).filter((header: Header) =>
-        this.showKeys.includes(header.key)
-      );
-    },
-  },
-  methods: {
-    getParticipant(token: string): string {
-      console.debug("getParticipant", token);
-      const participant = this.participants.find(
-        (participant: ParticipantModel) => participant.token === token
-      );
-      return participant
-        ? `${participant.participant_info.firstname} ${participant.participant_info.lastname}`
-        : token;
-    },
+    });
+
+    const headersFromKeys = computed<Header[]>(() => {
+      return showKeys.value.map((key) => ({
+        title: key,
+        key: key,
+        align: "start",
+        sortable: true,
+      }));
+    });
+
+    return {
+      filteredRecords,
+      headersFromKeys,
+      tableOptions,
+      getParticipant,
+    };
   },
 });
 </script>
