@@ -1,12 +1,22 @@
 export const chartColors = [
+  "#2E8B57",
+  "#8F1616",
   "#A4A4A4",
   "#7468E8",
-  "#D0D0D0",
+  "#FF6347",
   "#9752CB",
   "#AC004B",
   "#3CB371",
   "#D3802B",
   "#DDBC42",
   "#313131",
-  "#8F1616",
+  "#D0D0D0",
+  "#4682B4",
+  "#32CD32",
+  "#FFD700",
+  "#8B0000",
+  "#00CED1",
+  "#9400D3",
+  "#FF4500",
+  "#1E90FF",
 ];
