@@ -1,10 +1,7 @@
 import { expect, describe, it, beforeEach } from "vitest";
-import {
-  filterResponses,
-  addExpiredEntries,
-  parseDataForAreaChart,
-  createTimelineFor,
-} from "./chartFunctions";
+import { createTimelineFor, createNumericChartData } from "./chartFunctions";
+import { parseDataForAreaChart } from "./yesno-charts";
+import { addExpiredEntries } from "./shared-chartFunctions";
 import { setActivePinia, createPinia } from "pinia";
 import {
   responses1,
@@ -13,19 +10,11 @@ import {
   surveyList1,
   questionList1,
   chartData1,
-  chartDataX,
+  chartDataYesNo,
+  chartDataNumerical,
+  chartDataFreeText,
 } from "../testData/chartFunctionsTestData";
-import { koordStore } from "../store";
-
-describe("testFilteredResponses", () => {
-  it("should return an empty array if no responses are provided", () => {
-    expect(filterResponses("test", [])).toEqual([]);
-  });
-
-  it("should return an array of FilteredResponses", () => {
-    expect(filterResponses("G01Q01HO", responses1)).toEqual(filteredResponses1);
-  });
-});
+import { useSurveyStore } from "../store/surveyStore";
 
 describe("testAddExpiredEntries", () => {
   beforeEach(() => {
@@ -37,9 +26,7 @@ describe("testAddExpiredEntries", () => {
   });
 
   it("should return an array of FilteredResponses enriched with expired resposes", () => {
-    expect(addExpiredEntries(filteredResponses1)).toEqual(
-      filteredResponsesWithExpired
-    );
+    expect(addExpiredEntries(filteredResponses1)).toEqual(filteredResponsesWithExpired);
   });
 });
 
@@ -48,27 +35,38 @@ describe("testParseDataForAreaChart", () => {
     expect(parseDataForAreaChart([])).toEqual([]);
   });
 
-  it("should return an array of FilteredResponses enriched with expired resposes", () => {
-    expect(parseDataForAreaChart(filteredResponsesWithExpired)).toEqual(
-      chartData1
-    );
+  it("should return an areaChartData", () => {
+    expect(parseDataForAreaChart(filteredResponsesWithExpired)).toEqual(chartData1);
   });
 });
 
 describe("testCreateTimelineFor", () => {
+  setActivePinia(createPinia());
+  const store = useSurveyStore();
+
   beforeEach(() => {
-    setActivePinia(createPinia());
-    const store = koordStore();
     store.updateSurveyList(surveyList1);
     store.updateQuestions(surveyList1[0].sid, questionList1);
     store.settings.expirationTime = 7;
     store.responses[123456] = responses1;
   });
   it("should return an empty array if no responses are provided", () => {
-    expect(createTimelineFor("test", 999999)).toEqual({ datasets: [] });
+    store.selectedSurveyID = 999999;
+    expect(createTimelineFor("test")).toEqual({ datasets: [] });
   });
 
-  it("should return an array of FilteredResponses enriched with expired resposes", () => {
-    expect(createTimelineFor("G01Q01HO", 123456)).toEqual(chartDataX);
+  it("should return an chartData for a YESNO chart", () => {
+    store.selectedSurveyID = 123456;
+    expect(createTimelineFor("G01Q01HO")).toEqual(chartDataYesNo);
+  });
+
+  it("should return an chartData for a Numerical chart", () => {
+    store.selectedSurveyID = 123456;
+    expect(createNumericChartData("G01Q04TEMP")).toEqual(chartDataNumerical);
+  });
+
+  it("should return an chartData for a Freetext chart", () => {
+    store.selectedSurveyID = 123456;
+    expect(createTimelineFor("G01Q05FREE")).toEqual(chartDataFreeText);
   });
 });
