@@ -1,34 +1,20 @@
 <template>
   <v-container>
+    <v-row>
+      <display-options :options="chartOptions" />
+    </v-row>
     <v-row class="py-4" style="max-height: 1200px">
-      <v-col cols="12" class="mb-5">
-        <Slider
-          v-model="responseRange"
-          :min="new Date('2023-01-01').getTime()"
-          :max="new Date('2023-12-31').getTime()"
-          :step="24 * 60 * 60 * 1000"
-          :format="tooltipFormater"
-          @change="reCalcTestData"
-        />
-      </v-col>
+      <v-btn text="Refresh" @click="reCalcTestData" />
       <v-col cols="12">
         <h2>REFERENCE</h2>
-        <line-chart
-          :data="testData"
-          style="max-height: 600px"
-          :options="areaChartOptions"
-        />
+        <line-chart :data="testData" style="max-height: 600px" :options="areaChartOptions" />
       </v-col>
     </v-row>
 
     <v-row class="py-4" style="max-height: 1200px">
       <v-col cols="12">
         <h2>TEST</h2>
-        <line-chart
-          :data="chartDataX"
-          style="max-height: 600px"
-          :options="areaChartOptions"
-        />
+        <line-chart :data="chartDataX" style="max-height: 600px" :options="areaChartOptions" />
       </v-col>
     </v-row>
   </v-container>
@@ -37,14 +23,10 @@
 <script lang="ts">
 import { defineComponent, onBeforeUnmount, ref } from "vue";
 import { Line as LineChart } from "vue-chartjs";
-import Slider from "@vueform/slider";
-import { chartData1 as series, chartDataX } from "./chartFunctionsTestData";
+import DisplayOptions from "../components/surveys/DisplayOptions.vue";
+import { chartData1 as series, chartDataYesNo } from "./chartFunctionsTestData";
 import { createTimelineFor } from "../helpers/chartFunctions";
-import {
-  surveyList1,
-  questionList1,
-  responses1,
-} from "../testData/chartFunctionsTestData";
+import { surveyList1, questionList1, responses1 } from "../testData/chartFunctionsTestData";
 import {
   Chart as ChartJS,
   CategoryScale,
@@ -57,44 +39,36 @@ import {
   TimeScale,
   TimeSeriesScale,
   Filler,
+  ChartData,
 } from "chart.js";
-import { koordStore } from "../store";
+import { useSurveyStore } from "../store/surveyStore";
 import { storeToRefs } from "pinia";
 import { areaChartOptions } from "../components/surveys/line-options";
-import { tooltipFormater } from "../helpers/slider";
+import { chartOptions } from "../components/surveys/options";
 
 import "chartjs-adapter-moment";
 
-ChartJS.register(
-  CategoryScale,
-  LinearScale,
-  PointElement,
-  LineElement,
-  TimeScale,
-  TimeSeriesScale,
-  Filler,
-  Title,
-  Tooltip,
-  Legend
-);
+ChartJS.register(CategoryScale, LinearScale, PointElement, LineElement, TimeScale, TimeSeriesScale, Filler, Title, Tooltip, Legend);
 
 export default defineComponent({
   name: "PidraKin",
   components: {
     LineChart,
-    Slider,
+    DisplayOptions,
   },
   setup() {
-    const store = koordStore();
+    const store = useSurveyStore();
     const { responseRange } = storeToRefs(store);
     store.updateSurveyList(surveyList1);
     store.updateQuestions(surveyList1[0].sid, questionList1);
     store.settings.expirationTime = 7;
     store.responses[123456] = responses1;
-    const testData = ref(createTimelineFor("G01Q01HO", 123456));
+    store.responseRange = [new Date("2023-02-10").getTime(), new Date("2023-05-31").getTime()];
+    store.selectedSurveyID = 123456;
+    const testData = ref(createTimelineFor("G01Q01HO") as ChartData<"line">);
 
     function reCalcTestData() {
-      testData.value = createTimelineFor("G01Q01HO", 123456);
+      testData.value = createTimelineFor("G01Q01HO") as ChartData<"line">;
     }
 
     onBeforeUnmount(() => {
@@ -105,9 +79,9 @@ export default defineComponent({
       testData,
       responseRange,
       series,
-      chartDataX,
+      chartDataX: chartDataYesNo,
       areaChartOptions,
-      tooltipFormater,
+      chartOptions,
       reCalcTestData,
     };
   },
