@@ -43,5 +43,11 @@ export interface responseCount {
 export interface FilteredResponse {
   token: string;
   time: Date;
-  value: string;
+  answer: string | Record<string, string>;
+}
+export interface HLResponse {
+  token: string;
+  time: Date;
+  lowValue: number;
+  highValue: number;
 }
