@@ -6,21 +6,16 @@
     <v-row>
       <p>
         Please authenticate using your
-        <span class="font-weight-bold"
-          >LimeSurvey
-          <a v-if="instance != ''" :href="instance">[{{ instance }}]</a></span
-        >
+        <span class="font-weight-bold">
+          LimeSurvey
+          <a v-if="instance != ''" :href="instance">[{{ instance }}]</a>
+        </span>
         log-in credentials.
       </p>
     </v-row>
     <v-row>
       <v-col cols="4">
-        <login
-          :disabled="authenticating"
-          @auth-before="setBusy"
-          @auth-fail="setFailed"
-          @auth-success="setSuccess"
-        />
+        <login :disabled="authenticating" @auth-before="setBusy" @auth-fail="setFailed" @auth-success="setSuccess" />
       </v-col>
     </v-row>
   </v-container>
@@ -28,56 +23,58 @@
 
 <script lang="ts">
 import Login from "../components/Login.vue";
-import { defineComponent } from "vue";
-import { mapState, mapActions } from "pinia";
-import { koordStore } from "../store";
+import { defineComponent, onBeforeMount, ref } from "vue";
+import { storeToRefs } from "pinia";
+import { useRouter } from "vue-router";
+import { useMainStore } from "../store/mainStore";
+import { useSurveyStore } from "../store/surveyStore";
 
 export default defineComponent({
   name: "LoginView",
-
   components: {
     Login,
   },
-
   props: {
     returnTo: {
       type: String,
       default: "/",
     },
   },
+  setup(props) {
+    const authenticating = ref(false);
+    const mainStore = useMainStore();
+    const surveyStore = useSurveyStore();
 
-  data() {
-    return {
-      authenticating: false,
-    };
-  },
+    const { isAuthenticated, instance } = storeToRefs(mainStore);
+    const router = useRouter();
 
-  computed: {
-    ...mapState(koordStore, ["isAuthenticated", "username", "instance"]),
-  },
-
-  mounted(): void {
-    this.$nextTick(() => {
-      if (this.isAuthenticated) {
-        this.$router.push("/");
+    onBeforeMount(() => {
+      if (isAuthenticated.value) {
+        router.push("/");
       }
     });
-  },
-  methods: {
-    ...mapActions(koordStore, ["refreshSurveys"]),
-    setBusy(): void {
-      this.authenticating = true;
-    },
 
-    setFailed(): void {
-      this.authenticating = false;
-    },
+    function setBusy(): void {
+      authenticating.value = true;
+    }
 
-    setSuccess(): void {
-      this.refreshSurveys();
-      const goTo = this.returnTo || "/";
-      this.$router.push(goTo);
-    },
+    function setFailed(): void {
+      authenticating.value = false;
+    }
+
+    function setSuccess(): void {
+      surveyStore.refreshSurveys();
+      router.push(props.returnTo);
+    }
+
+    return {
+      authenticating,
+      isAuthenticated,
+      instance,
+      setBusy,
+      setFailed,
+      setSuccess,
+    };
   },
 });
 </script>
