@@ -1,12 +1,17 @@
 import { defineConfig } from "vite";
 import vue from "@vitejs/plugin-vue";
 import vuetify from "vite-plugin-vuetify";
+import path from "path";
 
 // https://vitejs.dev/config/
 export default defineConfig({
-  base: "./",
   server: {
     host: "0.0.0.0",
+  },
+  resolve: {
+    alias: {
+      "@": path.resolve(__dirname, "./src"),
+    },
   },
   plugins: [vue(), vuetify({ autoImport: true })],
   build: {
